refactor(graph): split capability and lint helpers into their own modules
The graph command module had grown to 834 lines, mixing verb dispatch
with the record shape a capability run produces, the helpers that
execute and render it, and lint result validation and persistence. The
repository convention is that a module root reads as a table of
contents.

Running and reporting one capability moves to `capability`, and lint
result handling moves to `lint`, leaving the root at 498 lines of
imports, options, and the build, lint, test, and run verbs. No behaviour
changes; the moved items keep their signatures.
